main_project_backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
sqlalchemy>=2.0.0
alembic>=1.13.0
pymysql>=1.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.8.0
httpx>=0.27.0
langchain-skillkit>=0.3.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-deepseek>=0.1.0
langchain-mcp-adapters>=0.1.0
redis[asyncio]>=5.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
langchain-anthropic>=0.2.0
langchain-google-genai>=2.0.0
deepagents>=0.4.12
